requirements.txt
fastapi
uvicorn[standard]
python-multipart
tensorflow
tensorflow-stubs
numpy
Pillow